불러온 데이터가 없을 시 실행 안하는 조건 추가

+++ backend/data crawler/money.py
@@ -25,14 +25,20 @@ with closing(myutil.mydbconnect()) as conn:
             base_url = "https://www.koreaexim.go.kr/site/program/financial/exchangeJSON?"
             params = {"data": "AP01", "authkey": my_key, "searchdate": today.strftime("%Y%m%d")}
             response = requests.post(base_url, verify=False, data=params)
-
-            placeholder = ", ".join(["%s"] * (len(response.json()[0]) + 1))
-            stmt = "insert into `{table}` ({columns}) values ({values});".format(
-                table="money_exchange_from_koreaexim",
-                columns=",".join(response.json()[0].keys()) + ",time_fromapi",
-                values=placeholder,
-            )
-            cur.execute("TRUNCATE TABLE " + "money_exchange_from_koreaexim")
+            money_data = []
             for i in response.json():
-                cur.execute(stmt, list(i.values()) + [today])
-            conn.commit()
+                if i.pop("result") == 1:
+                    money_data.append(i)
+                else:
+                    continue
+            if money_data:
+                placeholder = ", ".join(["%s"] * (len(money_data[0]) + 1))
+                stmt = "insert into `{table}` ({columns}) values ({values});".format(
+                    table="money_exchange_from_koreaexim",
+                    columns=",".join(money_data[0].keys()) + ",time_fromapi",
+                    values=placeholder,
+                )
+                cur.execute("TRUNCATE TABLE " + "money_exchange_from_koreaexim")
+                for i in money_data:
+                    cur.execute(stmt, [myutil.if_float_conv(v) for v in i.values()] + [today])
+                conn.commit()
